Pipeline the engine
The renderering is now pipelined. The simulation and the rendering are
now fully decoupled. The simulation lives in the main thread, while the
rendering lives in its own thread. The simulation thread processes the
frame N while the render thread is processing the frame N-1.

For now resources are created on the simulation thread but that may
change in the future. Each thread has its own OpenGL context.

I decided to put the simulation on the main thread because the SDL
requires the event handler to run on the main thread.

This is a really really basic implementation and a lot has to be done in
order to have a robust, modern pipelined engine. Enhancements will be
made incrementally, as I profile the engine and get insights on what's
going on.