Rewrite stack allocator
Until now the double-bufferization of the frame packet was implemented
directly into the buffer pool. Two buffers of hard-coded size were
preallocated when the processed was started : the write buffer and the
read buffer. The current frame's packet was allocated on the write
buffer using the placement new operator. This was very unpractical to
use.

1. I was over-allocating based on the biggest packet I'd seen so far.
2. Because it was double buffered, the allocator couldn't be used for
things than survive the frame.
3. I had to recompile the engine when I wanted to change the frame
packets' size.

This changeset does two things. Firstly, it decouples the double
bufferization of the frame packet from the allocation strategy.
Secondly, it implements the concept of tagged-block allocators from
Naughty Dog's "Parallelizing the Naughty Dog Engine Using Fibers" talk
from GDC 2015.

The engine's wording hasn't changed though. The classes Buffer and
BufferPool are still there but have been reworked (because well, blocks
are still buffers). I don't yet know why but I couln't use 2 MiB huge
pages on my version of macOS so I use the default page size for now. I
know I was able to allocate 2 MiB huge pages on macOS in the past but
that'll be an issue for another day.